# backend/openapi/v1.json
{
  "components": {
    "schemas": {
      "Accepted": {
        "description": "What a call answers when the caller asked not to wait: the task it started.",
        "properties": {
          "job_id": {
            "description": "Follow it with `GET /jobs/{job_id}`, whose `result` holds the report.",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "type": "object"
      },
...
      "AlternativeDecision": {
        "description": "An alternative decision that was considered but not selected.",
        "properties": {
          "category": {
            "type": "string"
          },
          "confidence": {
            "format": "float",
            "type": "number"
          },
          "excluded_by": {
            "description": "Set when this rule matched but an exclusion vetoed it.",
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": "string"
          },
          "rule_name": {
            "type": "string"
          }
        },
        "required": [
          "rule_name",
          "category",
          "reason"
        ],
        "type": "object"
      },
...
      "CapacityForecast": {
        "description": "One destination folder, and the weight of what this plan sends to it.",
        "properties": {
          "fits": {
            "description": "`incoming_bytes` fits in `free_bytes`.",
            "type": "boolean"
          },
          "free_bytes": {
            "format": "int64",
            "type": "integer"
          },
          "incoming_bytes": {
            "description": "Bytes arriving from a *different* filesystem, which is the only traffic\nthat consumes space (see `fits`).",
            "format": "int64",
            "type": "integer"
          },
          "instance_id": {
            "type": "string"
          },
          "instance_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "minimum": 0,
            "type": "integer"
          },
          "path": {
            "type": "string"
          },
          "same_filesystem_bytes": {
            "description": "Bytes arriving from a folder that reports the same free space, and so\nalmost certainly sits on the same filesystem. A move there is a rename\nand costs nothing.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "instance_id",
          "path",
          "incoming_bytes",
          "same_filesystem_bytes",
          "free_bytes",
          "items",
          "fits"
        ],
        "type": "object"
      },
...
      "Decision": {
        "description": "A routing decision computed by the rule engine.",
        "properties": {
          "action": {
            "type": "string",
            "x-extensible-enum": [
              "move",
              "none",
              "skip"
            ]
          },
          "actor": {
            "description": "What set the decision off: `manual`, `schedule`, `webhook`, `api`, or\n`auto` for the simulation the automation runs after a sync.",
            "type": [
              "string",
              "null"
            ]
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/AlternativeDecision"
            },
            "type": "array"
          },
          "applied_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "confidence": {
            "description": "0.0 to 1.0, how many independent signals backed the winning rule.",
            "format": "float",
            "type": "number"
          },
          "current_root_folder": {
            "type": [
              "string",
              "null"
            ]
          },
          "decided_at": {
            "type": "string"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "instance_id": {
            "type": "string"
          },
          "instance_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_override": {
            "type": "boolean"
          },
          "matched_rule_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "matched_rule_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "media_id": {
            "type": "string"
          },
          "media_title": {
            "type": "string"
          },
          "media_type": {
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "reverted_at": {
            "description": "Set when the move was rolled back through `POST /decisions/revert`.",
            "type": [
              "string",
              "null"
            ]
          },
          "revertible": {
            "description": "Whether a revert may undo this move now: only the latest standing move\nof a title is, while the title is still where it put it and the folder\nit came from is still one of its instance's root folders.",
            "type": "boolean"
          },
          "simulation_id": {
            "description": "Groups every decision produced by the same simulation run.",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "x-extensible-enum": [
              "pending",
              "requested",
              "applied",
              "failed",
              "skipped"
            ]
          },
          "subject": {
            "description": "Who asked: an application's name, or the person a sign-in mode names.\nNull when nobody asked (the scheduler or a webhook), and under a mode\nthat names nobody. An application key reads its own name and null for\nanyone else.",
            "type": [
              "string",
              "null"
            ]
          },
          "superseded": {
            "description": "True once this decision was retired: by a newer simulation, an override\nset or removed, a revalidation that found it stale, or the removal of\nits media or its instance.",
            "type": "boolean"
          },
          "target_category": {
            "type": "string"
          },
          "target_root_folder": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "media_id",
          "media_title",
          "media_type",
          "instance_id",
          "target_category",
          "is_override",
          "reasons",
          "alternatives",
          "action",
          "status",
          "decided_at"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The envelope every refusal and failure answers.",
        "properties": {
          "answerable": {
            "description": "Beside `confirm`: whether this caller may send the name back.",
            "type": [
              "boolean",
              "null"
            ]
          },
          "confirm": {
            "description": "The guardrail that asks, for `confirmation_required`. Sending it back\nin `confirm` accepts that question and no other.",
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "description": "A stable code: `bad_request`, `unauthorized`, `forbidden`, `not_found`,\n`method_not_allowed`, `conflict`, `confirmation_required`,\n`payload_too_large`, `unsupported_media_type`, `too_many_requests`,\n`busy`, `reauthentication_required`, `passphrase_required`,\n`external_api_error` or `internal_error`. An open list.",
            "type": "string"
          },
          "includes": {
            "description": "Beside `confirm`: the other guardrails the question states, sent back\nin `confirm` with it to accept the question whole.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "message": {
            "description": "A sentence for a person, in the interface language. Never part of the\ncontract.",
            "type": "string"
          },
          "reason": {
            "description": "Why a `conflict` refuses, an open list. A job still running, with a\n`Retry-After` header: `apply_running`, `simulation_running`,\n`previews_full`, `sync_running`, `enrichment_running`, `backup_running`,\n`backup_just_taken`, `restore_staging`, `purge_running`. A state that\nstays until somebody changes it: `name_taken`, `in_use`,\n`already_listed`, `not_declared`, `category_mapped`,\n`proposals_replaced`, `dry_run`, `not_cancellable`, `key_pinned`,\n`only_credential`, `secret_unreadable`.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "detail": {
            "description": "What the task did, in the interface language.",
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "description": "The code the call would have answered had it waited, the envelope's\n`error`, for a task that failed on an error. Null otherwise, and for a\nfailure the task reports itself, such as every move of an apply\nrefused.",
            "type": [
              "string",
              "null"
            ]
          },
          "error_message": {
            "description": "Why the task failed, a sentence for a person.",
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "instance_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "description": "`sync`, `sync_all`, `enrich`, `simulate`, `preview` (a simulation that\nstores nothing), `apply`, `revert`, `backup`, `maintenance` or\n`scheduler`.",
            "type": "string"
          },
          "progress_current": {
            "format": "int64",
            "type": "integer"
          },
          "progress_total": {
            "format": "int64",
            "type": "integer"
          },
          "result": {
            "description": "What a finished task answered: the report the same call gives when the\ncaller waits, an `ApplyReport`, a `BatchApplyReport` or a `SyncReport`,\nor for a simulation its `SimulationResult`. A stored simulation's comes\nwithout `decisions`, which `GET /decisions?simulation_id=` lists, and a\npreview's keeps the ones it returned. An apply or a revert that failed\nkeeps the report of what it attempted. Null while the task runs, when\nany other task failed, for a kind no call starts, and in the list\nunless it asks for it."
          },
          "started_at": {
            "type": "string"
          },
          "status": {
            "description": "`running`, `success`, `failed`, or `cancelled` for an apply or a\nrevert somebody stopped.",
            "type": "string"
          },
          "subject": {
            "description": "Who asked: an application's name, or the person a sign-in mode names.\nAn application key reads its own name and null for anyone else.",
            "type": [
              "string",
              "null"
            ]
          },
          "trigger": {
            "description": "What set the task off: `manual`, `schedule`, `webhook`, `api`, or `auto`\nfor what the automation runs after a sync.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "trigger",
          "progress_current",
          "progress_total",
          "started_at"
        ],
        "type": "object"
      },
...
      "SimulationRequest": {
        "additionalProperties": false,
        "description": "Request to run a simulation.",
        "properties": {
          "instance_ids": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "max_returned": {
            "description": "Cap the payload size. The counters always reflect the full library.",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "media_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "persist": {
            "description": "Store the decisions so they can be applied. `false` is a pure preview.",
            "type": "boolean"
          },
          "persist_unchanged": {
            "description": "Also keep decisions for media already in the right place.",
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SimulationResult": {
        "description": "Simulation result summary.",
        "properties": {
          "already_correct": {
            "minimum": 0,
            "type": "integer"
          },
          "capacity": {
            "description": "What the plan would put on each destination, and whether it fits.\n\nOnly destinations that receive something appear, and only when the Arr\nreported a free-space figure for them.",
            "items": {
              "$ref": "#/components/schemas/CapacityForecast"
            },
            "type": "array"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array"
          },
          "elapsed_ms": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "excluded_by_rule": {
            "description": "How often a rule matched a title and one of its own exclusions set it\naside, over every title, rules below the winner included.",
            "minimum": 0,
            "type": "integer"
          },
          "moves_required": {
            "minimum": 0,
            "type": "integer"
          },
          "no_category_match": {
            "minimum": 0,
            "type": "integer"
          },
          "overrides_applied": {
            "minimum": 0,
            "type": "integer"
          },
          "returned": {
            "description": "How many decisions are included in `decisions` after truncation.",
            "minimum": 0,
            "type": "integer"
          },
          "simulation_id": {
            "type": "string"
          },
          "skipped_unmapped": {
            "description": "Matched a category with no root folder mapped on that instance.",
            "minimum": 0,
            "type": "integer"
          },
          "total_media": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "simulation_id",
          "total_media",
          "returned",
          "decisions",
          "moves_required",
          "already_correct",
          "no_category_match",
          "overrides_applied",
          "skipped_unmapped",
          "excluded_by_rule",
          "capacity",
          "elapsed_ms"
        ],
        "type": "object"
      },
...
      "SyncReport": {
        "description": "Result of syncing one instance.",
        "properties": {
          "error": {
            "description": "Set when the sync failed, so the caller of sync-all still gets one entry\nper instance instead of the failure silently vanishing from the report.",
            "type": [
              "string",
              "null"
            ]
          },
          "instance_id": {
            "type": "string"
          },
          "instance_name": {
            "type": "string"
          },
          "media": {
            "minimum": 0,
            "type": "integer"
          },
          "removed": {
            "description": "Media rows that disappeared upstream and were removed locally.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "root_folders": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "instance_id",
          "instance_name",
          "root_folders",
          "media",
          "removed"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      },
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "identifier": "GPL-3.0-only",
      "name": "GPL-3.0-only"
    },
    "title": "Routarr API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/instances/sync": {
      "post": {
        "description": "Answers once every instance is read, one report each. An instance that\nfails is reported with its `error` and stops none of the others. With\n`respond-async` it answers once its task has started, and that task holds\nthe reports when it ends. Each instance is read under a task of its own too.",
        "operationId": "sync_all",
        "parameters": [
          {
            "description": "`respond-async` answers 202 once the task has started, instead of its reports.",
            "in": "header",
            "name": "Prefer",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SyncReport"
                  },
                  "type": "array"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "The task has started, as `Prefer: respond-async` asked.",
            "headers": {
              "Location": {
                "description": "The task, under `/jobs`.",
                "schema": {
                  "type": "string"
                }
              },
              "Preference-Applied": {
                "description": "`respond-async`.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Seconds to wait before the first look at the task.",
                "schema": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Read every enabled Arr again",
        "tags": [
          "instances"
        ],
        "x-routarr-scope": "operate"
      }
    },
...
    "/jobs/{id}": {
      "get": {
        "operationId": "get_task",
        "parameters": [
          {
            "description": "The task's id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "One task",
        "tags": [
          "tasks"
        ],
        "x-routarr-scope": "read"
      }
    },
...
    "/simulate": {
      "post": {
        "description": "With `persist: true` the decisions are stored under the `simulation_id`\nthis answers, replacing the pending ones of an earlier run. Only one\npersisting simulation runs at a time: another answers 409 until it ends.\nA preview, `persist: false`, waits its turn, and answers 409 while four\nare already running or waiting.",
        "operationId": "simulate",
        "parameters": [
          {
            "description": "`respond-async` answers 202 once the task has started, instead of its report.",
            "in": "header",
            "name": "Prefer",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "The task has started, as `Prefer: respond-async` asked.",
            "headers": {
              "Location": {
                "description": "The task, under `/jobs`.",
                "schema": {
                  "type": "string"
                }
              },
              "Preference-Applied": {
                "description": "`respond-async`.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Seconds to wait before the first look at the task.",
                "schema": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Run the rules over the library",
        "tags": [
          "decisions"
        ],
        "x-routarr-scope": "operate"
      }
    },
...
  },
  "security": [
    {
      "api_key": []
    },
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}